Add concept of scratch directory, move DART and forecasts
In some HPCs, you get a separate scratch directory with a bigger allowance but where files
are regularly scrubbed up. I am adding the concept of a `scratch` directory to the experiments
to store wrfout and DART files and a future commit will include features to extract some variables
into other files, to create your forecast and analysis files.